refactored deprecated components and fixed tests
