+++ config/DeprecatedComponent.js
@@ -1,9 +1,9 @@
 // @flow
 import React from "react";
 
-const DeprecatedComponent = (Component: React.Component, text: string) => {
+const DeprecatedComponent = (Component: React$ComponentType<any>, text: string) => {
   let warned = false;
-  return props => {
+  return (props: Object) => {
     if (!warned) {
       warned = true;
       if (process.env.NODE_ENV !== "production") console.warn(`[Deprecation] ${text}`);
