+++ src/_deprecated/InputText/index.js
@@ -1,38 +1,6 @@
 // @flow
-import * as React from "react";
-
-import Label from "../Input/Label";
-import FieldFeedback from "../FieldFeedback";
+import InputText from "./InputText";
 import DeprecatedComponent from "../../../config/DeprecatedComponent";
-
-type Props = {
-  value: string,
-  onChange: (SyntheticInputEvent<HTMLInputElement>) => any,
-  label?: string,
-  error?: string,
-  help?: string,
-};
-
-const InputText = (props: Props) => (
-  <React.Fragment>
-    <Label label={props.label} valueEmpty={!!props.value}>
-      <input value={props.value} onChange={props.onChange} />
-    </Label>
-    <FieldFeedback error={props.error} help={props.help} />
-    <style jsx>{`
-      input {
-        box-sizing: border-box;
-        width: 100%;
-        border-radius: 3px;
-        background-color: white;
-        border: solid 1px ${props.error ? "#d02228" : "#bac7d5"};
-        font-size: 16px;
-        line-height: 1.25;
-        padding: 12px 16px;
-      }
-    `}</style>
-  </React.Fragment>
-);
 
 export default DeprecatedComponent(
   InputText,
